requirements.txt
slack-bolt==1.21.3
slack-sdk==3.33.5
flask==3.1.0
apscheduler==3.10.4
python-dotenv==1.0.1
requests==2.32.3
anthropic==0.42.0
SQLAlchemy==2.0.36
